Selections with zero results are prevented
If a (category) selection would result in zero items in the list, that
selection is not applied. Also, the mouse-over select visual feedback
is disabled in such case, telling users that zero-cound categories are
not clickable. There are some more details, such as multi-selection on
a zero-count category still works, since it may "add" items instead of
remove items from the list.